# backend/app/api/v1/pool.py
"""Read-only access to the **simulated** item banks (plan §8).

Additive surface so the UI can demonstrate the linear workflow end-to-end with
genuine simulated data when no real item-factory export is wired. Lists the pool
catalog and serves a selected bank by ``pool_id``. Reads fixtures; changes nothing.
"""

from __future__ import annotations

from collections import Counter

from fastapi import APIRouter, Depends, HTTPException, Query
from sqlalchemy.orm import Session

from app.core.db import get_db
from app.psychometrics import pools
from app.psychometrics.params import require_metric
from app.schemas.pool import (
    ItemExposure,
    PoolCatalog,
    PoolDocument,
    PoolExposure,
    PoolItem,
    PoolSummary,
)
from app.services import item_exposure

router = APIRouter(prefix="/pool", tags=["pool"])

_TAG_DIMENSIONS = ("KC", "Bloom", "TIMSS", "domain")


@router.get("/exposure", response_model=PoolExposure)
def get_pool_exposure(
    pool_id: str = Query(default=pools.DEFAULT_POOL_ID),
    db: Session = Depends(get_db),
) -> PoolExposure:
    """Cumulative longitudinal item exposure for a pool (published vs draft usage)."""
    if not pools.is_known(pool_id):
        raise HTTPException(status_code=404, detail=f"unknown pool_id {pool_id!r}")
    summary = item_exposure.exposure_summary(db, pool_id=pool_id)
    items = [
        ItemExposure(
            item_id=iid,
            published=e["published"],
            assembled=e["assembled"],
            total=e["total"],
            n_forms=e["n_forms"],
            last_used=e["last_used"].isoformat() if e["last_used"] else None,
        )
        for iid, e in sorted(summary.items())
    ]
    return PoolExposure(
        pool_id=pool_id,
        exposure_contexts=list(item_exposure.DEFAULT_EXPOSURE_CONTEXTS),
        items=items,
    )


@router.get("/catalog", response_model=PoolCatalog)
def get_pool_catalog() -> PoolCatalog:
    """List the selectable simulated banks (for the pool selector)."""
    summaries: list[PoolSummary] = []
    for entry in pools.catalog():
        doc = pools.load_document_by_id(entry.pool_id)
        raw = doc["items"]
        domains = sorted(
            {it["tags"]["domain"] for it in raw if "domain" in it.get("tags", {})}
        )
        summaries.append(
            PoolSummary(
                pool_id=entry.pool_id,
                title=entry.title,
                description=entry.description,
                model=(
                    "none (uncalibrated)"
                    if doc.get("field_pool")
                    else doc.get("metric", {}).get("model", "2PL")
                ),
                simulated=bool(doc.get("simulated", False)),
                n_items=len(raw),
                n_3pl=sum(1 for it in raw if it.get("c", 0.0) > 0),
                domains=domains,
            )
        )
    return PoolCatalog(default_pool_id=pools.DEFAULT_POOL_ID, pools=summaries)


@router.get("/items", response_model=PoolDocument)
def get_pool_items(
    pool_id: str = Query(default=pools.DEFAULT_POOL_ID),
) -> PoolDocument:
    """Return a simulated bank with params, tags, and synthetic content."""
    if not pools.is_known(pool_id):
        raise HTTPException(status_code=404, detail=f"unknown pool_id {pool_id!r}")
    doc = pools.load_document_by_id(pool_id)
    raw_items = doc["items"]

    # Content-only field-study pools: items carry tags/content but NO parameters —
    # served with a/d/b None (honestly absent) so the editor's availability hints
    # and the viewer work; nothing psychometric can be computed from this document.
    if pools.is_field_pool(pool_id):
        items = [
            PoolItem(
                item_id=it["item_id"],
                tags=it.get("tags", {}),
                enemy_of=it.get("enemy_of", []),
                calibrated_anchor=bool(it.get("calibrated", False)),
                stem=it.get("stem"),
                options=it.get("options", []),
                answer_key=it.get("answer_key"),
            )
            for it in raw_items
        ]
        tag_summary_f: dict[str, dict[str, int]] = {}
        dims = sorted({k for it in raw_items for k in it.get("tags", {})})
        for dim in dims:
            counts = Counter(
                it["tags"][dim] for it in raw_items if dim in it.get("tags", {})
            )
            if counts:
                tag_summary_f[dim] = dict(sorted(counts.items()))
        return PoolDocument(
            pool_id=pool_id,
            simulated=False,
            provenance=doc.get("provenance"),
            model="none (uncalibrated)",
            scaling_d=None,
            form="none",
            kind="field",
            pool_kind="field",
            n_items=len(items),
            tag_summary=tag_summary_f,
            items=items,
        )

    # Enforce the metric contract (no silent default; raises on undeclared metric).
    metric = require_metric(doc.get("metric"), where=f"pool {pool_id!r}")

    def _b(it: dict) -> float:
        if metric.form == "slope_intercept":
            return -float(it["d"]) / float(it["a"])
        return float(it["b"])

    def _d(it: dict) -> float:
        if metric.form == "slope_intercept":
            return float(it["d"])
        return -float(it["a"]) * float(it["b"])

    items = [
        PoolItem(
            item_id=it["item_id"],
            a=it["a"],
            d=_d(it),
            c=it.get("c", 0.0),
            u=it.get("u", 1.0),
            b=_b(it),
            scaling_d=metric.scaling_d,
            tags=it.get("tags", {}),
            enemy_of=it.get("enemy_of", []),
            se_a=it.get("se_a"),
            se_d=it.get("se_d"),
            se_b=it.get("se_b"),
            stem=it.get("stem"),
            options=it.get("options", []),
            answer_key=it.get("answer_key"),
        )
        for it in raw_items
    ]

    tag_summary: dict[str, dict[str, int]] = {}
    extra_dims = sorted(
        {k for it in raw_items for k in it.get("tags", {})} - set(_TAG_DIMENSIONS)
    )
    for dim in (*_TAG_DIMENSIONS, *extra_dims):
        counts = Counter(
            it["tags"][dim] for it in raw_items if dim in it.get("tags", {})
        )
        if counts:
            tag_summary[dim] = dict(sorted(counts.items()))

    return PoolDocument(
        pool_id=pool_id,
        simulated=bool(doc.get("simulated", False)),
        provenance=doc.get("provenance"),
        model=doc.get("metric", {}).get("model", "2PL"),
        scaling_d=metric.scaling_d,
        form=metric.form,
        kind=metric.kind,
        n_items=len(items),
        tag_summary=tag_summary,
        items=items,
    )