# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tests Platform API",
    "description": "Assembly + administration engine for a large testing program.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/pool/exposure": {
      "get": {
        "tags": [
          "pool"
        ],
        "summary": "Get Pool Exposure",
        "description": "Cumulative longitudinal item exposure for a pool (published vs draft usage).",
        "operationId": "get_pool_exposure",
        "parameters": [
          {
            "name": "pool_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "small_2pl",
              "title": "Pool Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoolExposure"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/pool/items": {
      "get": {
        "tags": [
          "pool"
        ],
        "summary": "Get Pool Items",
        "description": "Return a simulated bank with params, tags, and synthetic content.",
        "operationId": "get_pool_items",
        "parameters": [
          {
            "name": "pool_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "small_2pl",
              "title": "Pool Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoolDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/preview/start": {
      "post": {
        "tags": [
          "preview"
        ],
        "summary": "Start Preview",
        "operationId": "start_preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewStep"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/preview/respond": {
      "post": {
        "tags": [
          "preview"
        ],
        "summary": "Respond Preview",
        "operationId": "respond_preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRespondRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewStep"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/preview/score": {
      "post": {
        "tags": [
          "preview"
        ],
        "summary": "Score Preview",
        "operationId": "score_preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemExposure": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "published": {
            "type": "integer",
            "title": "Published",
            "default": 0
          },
          "assembled": {
            "type": "integer",
            "title": "Assembled",
            "default": 0
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "n_forms": {
            "type": "integer",
            "title": "N Forms",
            "default": 0
          },
          "last_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used"
          }
        },
        "type": "object",
        "required": [
          "item_id"
        ],
        "title": "ItemExposure",
        "description": "Cumulative longitudinal usage of one item across assemblies over time."
      },
...
      "Navigation": {
        "properties": {
          "can_review": {
            "type": "boolean",
            "title": "Can Review",
            "default": false
          },
          "can_skip": {
            "type": "boolean",
            "title": "Can Skip",
            "default": false
          },
          "can_navigate_back": {
            "type": "boolean",
            "title": "Can Navigate Back",
            "default": false
          },
          "fixed_length": {
            "type": "boolean",
            "title": "Fixed Length",
            "default": true
          },
          "total_items": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Items"
          }
        },
        "type": "object",
        "title": "Navigation",
        "description": "Navigation capabilities exposed to the test-taker / Sessions module."
      },
      "NextAction": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "present",
              "complete"
            ],
            "title": "Kind"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "navigation": {
            "$ref": "#/components/schemas/Navigation"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "navigation"
        ],
        "title": "NextAction",
        "description": "What the engine should do next: present item(s)/module, or complete."
      },
...
      "PoolDocument": {
        "properties": {
          "pool_id": {
            "type": "string",
            "title": "Pool Id"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated"
          },
          "provenance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "scaling_d": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scaling D"
          },
          "form": {
            "type": "string",
            "title": "Form"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "n_items": {
            "type": "integer",
            "title": "N Items"
          },
          "pool_kind": {
            "type": "string",
            "title": "Pool Kind",
            "default": "parametric"
          },
          "tag_summary": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "type": "object",
            "title": "Tag Summary"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PoolItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "pool_id",
          "simulated",
          "model",
          "form",
          "kind",
          "n_items",
          "tag_summary",
          "items"
        ],
        "title": "PoolDocument"
      },
      "PoolExposure": {
        "properties": {
          "pool_id": {
            "type": "string",
            "title": "Pool Id"
          },
          "exposure_contexts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exposure Contexts"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemExposure"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "pool_id",
          "exposure_contexts",
          "items"
        ],
        "title": "PoolExposure"
      },
      "PoolItem": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "a": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "A"
          },
          "d": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "D"
          },
          "c": {
            "type": "number",
            "title": "C",
            "default": 0.0
          },
          "u": {
            "type": "number",
            "title": "U",
            "default": 1.0
          },
          "b": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "B"
          },
          "scaling_d": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scaling D"
          },
          "calibrated_anchor": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibrated Anchor"
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Tags",
            "default": {}
          },
          "enemy_of": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Enemy Of",
            "default": []
          },
          "se_a": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Se A"
          },
          "se_d": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Se D"
          },
          "se_b": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Se B"
          },
          "stem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stem"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/PoolItemOption"
            },
            "type": "array",
            "title": "Options",
            "default": []
          },
          "answer_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Key"
          }
        },
        "type": "object",
        "required": [
          "item_id"
        ],
        "title": "PoolItem"
      },
      "PoolItemOption": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "key",
          "text"
        ],
        "title": "PoolItemOption"
      },
...
      "PreviewRespondRequest": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/SessionState"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          }
        },
        "type": "object",
        "required": [
          "state",
          "correct"
        ],
        "title": "PreviewRespondRequest",
        "description": "Record one response against the carried-back session state."
      },
      "PreviewScoreRequest": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/SessionState"
          }
        },
        "type": "object",
        "required": [
          "state"
        ],
        "title": "PreviewScoreRequest",
        "description": "Score the carried-back session state."
      },
      "PreviewStartRequest": {
        "properties": {
          "blueprint_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blueprint Id"
          },
          "form_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Id"
          },
          "pool_id": {
            "type": "string",
            "title": "Pool Id",
            "default": "small_2pl"
          },
          "assembly_strategy": {
            "type": "string",
            "title": "Assembly Strategy",
            "default": "mip"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "title": "PreviewStartRequest",
        "description": "Start a dry-run from a stored blueprint (assemble now) or a stored form."
      },
      "PreviewStep": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/SessionState"
          },
          "next_action": {
            "$ref": "#/components/schemas/NextAction"
          },
          "termination": {
            "$ref": "#/components/schemas/TerminationDecision"
          }
        },
        "type": "object",
        "required": [
          "state",
          "next_action",
          "termination"
        ],
        "title": "PreviewStep",
        "description": "One step of the walkthrough: the new state + what to do next.\n\n``next_action.navigation`` carries the strategy's capabilities, so no separate\ncapabilities call is needed."
      },
...
      "ScoreResult": {
        "properties": {
          "theta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theta"
          },
          "standard_error": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standard Error"
          },
          "scale": {
            "type": "string",
            "title": "Scale",
            "default": "canonical"
          },
          "detail": {
            "additionalProperties": true,
            "type": "object",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "ScoreResult",
        "description": "Result of scoring a session on the canonical theta metric.\n\nAll theta values are normalized through ``psychometrics/`` (the single source\nof truth = the mirt scoring service). Mind the D-scaling mismatch."
      },
      "SessionState": {
        "properties": {
          "model_type": {
            "type": "string",
            "title": "Model Type"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "position": {
            "type": "integer",
            "title": "Position",
            "default": 0
          },
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "default": false
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "model_type",
          "session_id"
        ],
        "title": "SessionState",
        "description": "Opaque-to-the-core administration state for one in-flight session.\n\nThe engine core treats this as a black box and only passes it back to the\nowning strategy. Concrete strategies refine ``data`` with their own typed\npayloads in Phase 1+."
      },
...
      "TerminationDecision": {
        "properties": {
          "complete": {
            "type": "boolean",
            "title": "Complete"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "complete"
        ],
        "title": "TerminationDecision",
        "description": "Whether the session is complete and why."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}